update (frontend): basic tooling for development
I setup a basic common webpack config that will be merged with another depending on the environment.
Only the development config is available right now.
Babel is used for both eslint and transpiling the code. Babel is well supported and allows the use of react-refresh for rapid development.

normalize.css and PostCSS are used to minimize browser differences.

The combination of @loadable/component and webpack's import allow for code splitting in a simple manner